+++ example/config/deploy/production.rb
@@ -1,0 +1,1 @@
+server "do1", roles: %w(web app db)